efi: authenticate fit capsules
UEFI authenticated capsules are supported only when they are in raw format.
Extend the support for authentication to UEFI capsules in the FIT format.

We move the authentication code out of the raw format function into a new
efi_firmware_capsule_authenticate() function. The special case for the FMP
header coming from edk2 tools is preserved. There is no functional change
for capsules in raw format.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>
Cc: Heinrich Schuchardt <xypron.glpk@gmx.de>